requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
h5py>=3.8.0
pyarrow>=12.0.0

# Hyperparameter Optimization
optuna>=3.2.0
ray[tune]>=2.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
wandb>=0.15.0

# Jupyter and Development
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Configuration and Logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0

# Scientific Computing
biopython>=1.81
biopandas>=0.2.7

# Model Interpretability
shap>=0.42.0
lime>=0.2.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Utilities
click>=8.1.0
rich>=13.4.0
pathlib2>=2.3.7
